requirements.txt
azure.storage.blob == 12.2.0
PyPDF2 == 1.26.0
pathlib == 1.0.1